pr/backend/requirements.txt
# SAFEE Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

# Firebase
firebase-admin==6.5.0

# LangChain
langchain-core==0.3.0
langchain-mistralai==0.2.0

# ML Models
sentence-transformers==3.0.1
faiss-cpu==1.12.0
transformers==4.44.0
torch>=2.0.0

# Code quality tools (for tier-1 tests)
flake8==7.1.1
pytest==8.3.2

# GitHub Integration
PyGithub==2.3.0
